Split the manageable-server concern and centralize login
- ManageableServers splits into SetsManageableServers (records the cache, used
  by the picker/dashboard) and RequiresManageableServer (gates a controller).
  The latter applies before_action :require_manageable_server on include, so a
  server-scoped controller can't forget to authorize.
- require_login moves to ApplicationController (public controllers skip it), so
  it always runs before the concern's guard and login can't be forgotten either.
- Collapse ServersController's inline Discord-error rescues into rescue_from;
  the dashboard now handles token expiry / Discord-down the same way the picker
  does (re-auth / error state) instead of redirecting.
- Document that re-auth is localized to the guild-fetch seam, with a note to
  extract a DiscordReauth concern + return-to flow if a second token-using page
  ever appears.
