+++ app/controllers/pages_controller.rb
@@ -1,4 +1,6 @@
 class PagesController < ApplicationController
+  skip_before_action :require_login
+
   def home
     return redirect_to servers_path if current_user
