+++ app/controllers/sessions_controller.rb
@@ -1,4 +1,6 @@
 class SessionsController < ApplicationController
+  skip_before_action :require_login
+
   def create
     auth = request.env["omniauth.auth"]
     user = User.from_omniauth(auth)
