+++ docs/architecture.md
@@ -73,15 +73,27 @@ and the web app, so a given mutation is written once and called from both.
 - **Booleans from forms.** Let ActiveRecord cast checkbox values — assign the raw
   param to the model attribute (`"1"`/`"0"` → `true`/`false`) rather than coercing
   in the controller.
-- **Server-scoped authorization** lives in the `ManageableServers` concern, not in
-  individual controllers. We don't persist which servers a user manages (it's a
-  Discord fact, fetched per the metadata-sync design, not a DB relationship), so
-  authorization is cached in the signed session: the picker and dashboard call
-  `remember_manageable_servers` after proving manageability via Discord, and every
-  server-scoped controller gates its actions with `before_action
-  :require_manageable_server` (which also sets `@server_configuration`). That check
-  reads the session set instead of re-hitting Discord's rate-limited guild-list
-  endpoint. It is not spoofable — the session is server-signed.
+- **Server-scoped authorization** lives in two concerns, not in individual
+  controllers. We don't persist which servers a user manages (it's a Discord fact,
+  fetched per the metadata-sync design, not a DB relationship), so authorization is
+  cached in the signed session. `SetsManageableServers` (included by the picker and
+  dashboard) exposes `remember_manageable_servers`, called after proving
+  manageability via Discord. `RequiresManageableServer` (included by every
+  server-scoped controller) **applies `before_action :require_manageable_server`
+  on include** — so a controller can't forget to authorize — which checks the
+  cached set and sets `@server_configuration`. The check reads the session instead
+  of re-hitting Discord's rate-limited guild-list endpoint, and is not spoofable
+  (server-signed session).
+- **Re-authentication is localized to the guild-fetch seam.** The user's Discord
+  token is used in exactly one place — `Discord::UserGuilds.call`, from
+  `ServersController` (picker + dashboard) — so token-expiry handling (`rescue_from
+  Discord::UserGuilds::Unauthorized` → re-auth, `::Error` → error state) lives
+  there, not in `ApplicationController`. Everything else reads our own DB and
+  authorizes from the session cache, so an expired token elsewhere is harmless;
+  stale-cache cases redirect to the picker, which is where re-auth runs. **If a
+  second page ever fetches from Discord with the user token, revisit this**: extract
+  a shared `DiscordReauth` concern and add return-to-URL plumbing (so re-auth sends
+  the user back where they were) rather than always landing on the picker.
 - **Turbo Stream responses live in a view template**, not in controller helpers —
   the standard Rails `action.turbo_stream.erb` (e.g.
   `app/views/servers/plugins/update.turbo_stream.erb`). The controller sets its
